(chocolatey-community#58) Speed up Get-RemoteChecksum
The current Get_RemoteChecksum implementation uses Invoke-WebRequest
which can be slow on larger files. This results in a long checksumming
process during automated packaging. By switching to the .NET webclient,
while losing the "interactive" download update, we significantly increase
the speed of the download. For example, on a 150 mb file, it went from minutes
to download to seconds.

Added modified date in this commit